+++ CHANGELOG.md
@@ -8,6 +8,13 @@ View all releases at: https://github.com/trimble-oss/modus-web-components/releas
 
 ### Added
 
+- Added `backgroundColor`and `borderRadius` props for `modus-card` component in order to customize `background-color` and `border-radius` 
+- Added `showCardBorder` and `showShadowHover` flags props for `modus-card` component in order to display or not border for card component and shadow box when component is hovered.
+
+## 0.1.22 - 2022-11-04
+
+### Added
+
 - Added a theme toggler to the toolbar in storybook pages to switch between `light` and `dark` themes
 
 ## 0.1.22 - 2022-10-18